multiple improvements and refactors regarding utility functions to read http responses data and create history items based on it

+++ cmd/describe_history.go
@@ -22,7 +22,7 @@ var describeHistoryCmd = &cobra.Command{
 			fmt.Println("An ID needs to be provided")
 			os.Exit(0)
 		}
-		history, err := db.Connection.GetHistory(describeHistoryID)
+		history, err := db.Connection.GetHistory(uint(describeHistoryID))
 		if err != nil {
 			log.Panic().Err(err).Msg("Could not find a issue with the provided ID")
 		}
